customers som har gjort beställningar får namn och personnummer = null, men är kvar i tabellen. andra customers tas bort från tabellen lite uppdatering i customer-cards

+++ src/main/java/com/example/webbshopbackend1/Controllers/CustomerHTMLController.java
@@ -1,7 +1,9 @@
 package com.example.webbshopbackend1.Controllers;
 
 import com.example.webbshopbackend1.Models.Customer;
+import com.example.webbshopbackend1.Models.Orders;
 import com.example.webbshopbackend1.Repos.CustomerRepo;
+import com.example.webbshopbackend1.Repos.OrderRepo;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -12,9 +14,11 @@ import java.util.List;
 @RequestMapping("/customerHTML")
 public class CustomerHTMLController {
     private final CustomerRepo customerRepo;
+    private final OrderRepo orderRepo;
 
-    public CustomerHTMLController(CustomerRepo customerRepo) {
+    public CustomerHTMLController(CustomerRepo customerRepo, OrderRepo orderRepo) {
         this.customerRepo = customerRepo;
+        this.orderRepo = orderRepo;
     }
 
     @RequestMapping("/getAll")
@@ -31,23 +35,46 @@ public class CustomerHTMLController {
     public String getById(@PathVariable Long id, Model model) {
 
         Customer customer = customerRepo.findById(id).orElse(null);
-        model.addAttribute("fullname", customer.getName());
-        model.addAttribute("ssn", customer.getSocialSecurityNumber());
-        model.addAttribute("id", customer.getId());
-        model.addAttribute("nameTitle", "Full name");
-        model.addAttribute("ssnTitle", "Social security number");
-        model.addAttribute("idTitle", "Customer ID");
-        model.addAttribute("deleted", false);
+        int orders = orderRepo.findByCustomerId(id).size();
 
+        if (customer.getName()==null){
+            model.addAttribute("id", customer.getId());
+            model.addAttribute("idTitle", "Former Customer ID");
+            model.addAttribute("orderTitle", "Number of orders");
+            model.addAttribute("orders", orders);
+            model.addAttribute("deleted", true);
+        }else {
+            model.addAttribute("fullname", customer.getName());
+            model.addAttribute("ssn", customer.getSocialSecurityNumber());
+            model.addAttribute("id", customer.getId());
+            model.addAttribute("orders", orders);
+            model.addAttribute("nameTitle", "Full name");
+            model.addAttribute("ssnTitle", "Social security number");
+            model.addAttribute("idTitle", "Customer ID");
+            model.addAttribute("orderTitle", "Number of orders");
+            model.addAttribute("deleted", false);
+        }
         return "one-customer.html";
     }
 
     @RequestMapping("/deleteById/{id}")
     public String delete(@PathVariable Long id, Model model) {
-        customerRepo.deleteById(id);
-        model.addAttribute("message", "CUSTOMER DELETED!");
-        model.addAttribute("deleted", true);
-        return "one-customer.html";
+        int orders = orderRepo.findByCustomerId(id).size(); //skapa en int som motsvarar antalet ordrar kunden gjort
+        if (orders <1) {
+            customerRepo.deleteById(id);
+        }else {
+            Customer customer = customerRepo.findById(id).orElse(null);
+            customer.setName(null);
+            customer.setSocialSecurityNumber(null);
+            customerRepo.save(customer);
+            model.addAttribute("id", customer.getId());
+            model.addAttribute("idTitle", "Former Customer ID");
+            model.addAttribute("orderTitle", "Number of orders");
+            model.addAttribute("orders", orders);
+        }
+            model.addAttribute("message", "CUSTOMER DELETED!");
+            model.addAttribute("deleted", true);
+            return "one-customer.html";
     }
 
     @PostMapping("/save")
